Refactor : 디렉토리 리팩토링
디렉토리 구조 및 이름을 일관성있게 리팩토링하였습니다.
