+++ frontend/src/components/Main/ContentsSection/ServerJoinSection.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useState } from 'react';
 import styled from 'styled-components';
-import { fetchData } from '../../utils/fetchMethods';
-import { MainStoreContext } from './MainStore';
+import { fetchData } from '../../../utils/fetchMethods';
+import { MainStoreContext } from '../MainStore';
 
 const Container = styled.div`
   width: 100%;
