+++ frontend/src/components/Main/Server/Modal/QuitServerModal.tsx
@@ -1,8 +1,8 @@
 import React, { useContext, useState } from 'react';
 import styled from 'styled-components';
 
-import { MainStoreContext } from '../MainStore';
-import { deleteApi } from '../../../utils/fetchMethods';
+import { MainStoreContext } from '../../MainStore';
+import { deleteApi } from '../../../../utils/fetchMethods';
 
 const Container = styled.form`
   width: 90%;
